Initial commit
Was in a process of how to implement, Data access and the and the changes for that data. The objective is, that the application needs to work with any type of data storage, let it be text file, json, or a relational database. I was thinking on how should I plan my interfaces so that the aformentioned objective is met. My findings are: I think that the current implementation should work for Json access and Write, but my concerns are for the usage. The JsonWriter implementation needs a dependency for JsonDataProvider beacuse I need to deserialize the whole json file to add the Transaction, then serialize it, and then Write it all back to the original file. Beacuse of this, it violates several SOLID principle, so another approach is required. For a database implementation, I tought about how would it work with the current interfaces. It would return All records within the table using the Read implementation, for the other CRUD operations it would need their respective interfaces.