DataAccessImplementation - Json variant progress
Was thinking about the previous objective, that how should I plan and make my Interfaces, so that it'll be usable with any type of Data Access. My findings are: The previous Json implementation attempt was not sufficent, beacuse it violated multiple SOLID principles, instead I tried applying what I learned about Interface segregation, and it seems that this is the right direction. I made multiple Interfaces, each covering a CRUD operation, and then all of those Interfaces will be implemented by a singular Interface's consumer. With this approach, I was able to group all necessary function, which will be needed in future Data Access implementations, for example a Data Base implementation. Then I was thinking about how will it work with dependency injection, considering one Interface will have multiple implementation. In Program.cs I added the current ITransactionCrud implementation to services, with this approach I'll be able to choose which Data Access implementation I want to use in the future, without changing the majority of the code functionality. All this, is thanks to the depedency injection container.