DataAccessImplementation - Refactoring of interfaces, working json variant
After the discussion in the previous Pull Request, I revised the interfaces according the Repository tutorial. Necessary functionality are now grouped in a single interface, which will be implemented similarly like the previous attempt, only this is a more concise, less error prone solution.  I deemed the other parts of the tutorial  unnecesary, beacuse this is a simple use case. All of Respository pattern was not implemented, I only used the idea of a singular interface for a model, I didn't add another layer of abstraction like in the tutorial, which would be usefull if other models would be present, with the usage of a marker interface it would guarantee the correct implementation through a generic IRepository interface. It was all to avoid overcomplication, now it should be aligned with SOLID principles, and also with common sense. I tried out this solution's persistence functionality, with a simple GET, and POST endpoint, in the UI as well as in postman. Next step in DataAcces would be to implement, a Data base variant, or any other variant which would be also injected in the container, and would work similarly as the json variant, only with slight adjustments. After that, all Repository implementation will be interchangeable, and the original objective of this feature is completed.