UI Development - Transaction page implemented
This is a rather lenghty change, so I'll dissect it layer by layer: In the controller I added the necessary endpoints according to user story. Added a central view for Transactions, where I loaded in the add, modify forms and all previously persisted Transactions as partial views. This decision was motivated by the user story, which stated that the user should be able to see the previously added transactions, and can interact with them, meaning they can delete or modify them, and also should be able to add a new transaction in the same page. This was implemented with a twist, which is questionable, but it merits a discussion, I'll elaborate later. This was the controller part. Minor changes were made to the JsonRepository implementation, the re-read part was already performed by the Transactions property, so I deemed it unnecessary. The View part is where the magic happens. The user story states, that the user should be able to return to each central page in our case the Transaction page and to be implemented Dashboard page, using the hamburger menu. This menu is in the Layout page which is the template for all the pages that will be rendered within in it, meaning all central pages will have its' hamburger menu and the user can switch back and forth between them. Next is the Transaction page. This is where I rendered in the partial views, with the aformentioned twist. The problem was, that I needed to display the validation error messages for a partial view, after the user modified or wanted to save a new Transaction. The functionality of the Transaction page was that in the beginning the user sees all the previously entered transactions, and the new Transaction form. If the user wanted to modify a transation, and the validation found an error, the jquery part where the page is loaded in, didn't know if there was an error and loaded in the page, hiding the modify form. To solve this issue, I relied on ViewDataDictionary which contained the ModelState errors, provided by the validation endpoint. Here comes the twist part. I needed additional information, about which validation endpoint is called. I got this information, by passing the validation endpoint name through viewbag. Now in the document ready jquery function, only functions normaly, meaning hides modify form, clear the contents of new Transaction form, if there is no ModelState errors.  Using the actionName variable I can control which partial view can be loaded in, in case of Modelstate error. Here's an example: If there is an error the modify form I wont load in the add form, until the errors are resolved, beacuse the actionName is set to Modify which prevents the add form to load in, this applies to the modify form as well. The validation errors are displayed in the form's container, until there are no errors left. The partial views' functionality are pretty straight forward. In the previously added Transactions partial view, users can modify and delete a transaction. These are achived using jquery functions, which are invoked by onclick, and the button in question is passed to the jquery function. If a user wanted to delete a Transaction, it'll prompt a pop up, to confirm the deletion by clicking on the confirm button. The modify button is where, I load in the modify form partial view, instead of the add form. The modify and add forms are validated by their respective validation endpoints, and display if there's an error.