WIP: Dashboard implementation
This one gave me a bit of a headache, for a while I couldn't come up with a solution how to pass usable data to a Chart and display it on the dashboard page. I looked at how I stored data, specifically Dates, and after that I came up with a solution to integrate it to Chart.js and display it, the way the user story intended. The code itself reflects the current best solution what I could come up with, and I didn't apply (yet) the best practices to get a cleaner iplementation, for example DRY, there are a lot of instances where I repeat the same functionality, which can be used as a method, or can be a part of a class, a helper one perhaps. The last enpoint which is called "Average contribution amount by month and year" I couldn't understand what it entails or how should I implement it. It merits a discussion in the PR.